requirements-test.txt
pytest~=1.3.0
pytest-flask~=1.3.0
pytest-flask-sqlalchemy~=1.1.0
pytest-cov~=6.1.1